requirements.txt
streamlit
google-genai
python-docx
python-pptx
pandas
openpyxl
pypdf
XlsxWriter
pymupdf
python-calamine